Added pipeline features to LimelightProcessing.java, now we can switch pipelines and control apriltags from there Added docs to LimelightProcessing.java and TargetInfo Added a verify movement function to turret class, which ensures movement is within range, allows us to wrap encoder positions
In order to retain full encoder reliability, we will need to run the new class PreMatchHoming.java, which literally just moves the motor until we hit a limit switch then resets encoder to 0. If we run this every time, we should be able to wrap around nicely but who knows. Cleaned up various files as well.