Added botPose3D calculations to LimelightProcessing.java Maybe this returns the right numbers? IDK, i'm not sure if i'm accessing the right YAW variable from pedropathing/pinpoint
Added distance calculations to TargetInfo. This is solely based on area of the TargetInfo (gotten by LimelightProcessing.java), and a scale factor. We need to determine this scale factor

Added AprilTagRangeCalculator to do just that. Based on brogan pratt video, should give us all data needed to plot graph and get a scale factor to plug into our scale variable